+++ src/styles/user.css
@@ -1,22 +1,3 @@
-/*uyen-transfusions*/
-.modal-backdrop {
-  position: fixed;
-  top: 0; left: 0; right: 0; bottom: 0;
-  background: rgba(0, 0, 0, 0.1); 
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 9999;
-}
-
-.modal-content {
-  background: rgb(255, 255, 255);
-  padding: 20px;
-  border-radius: 8px;
-  width: 350px;
-  box-shadow: 0 4px 10px rgba(0,0,0,0.25);
-}
-
 /* Giao diện tổng thể */
 body {
   font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
